feat: add Safari header toggle and improve the style of attribution
Added a new Safari Header switch option so users can control whether or not to display a Safari style header.
Simplified the logic of showAttribution and changed the style to tailwindcss.
The type of delayTimer was changed from <NodeJS.Timeout> to < NodeJS.Timeout |null>, this may fix a memory leak caused by not clearing the timer when the component was uninstalled.